# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AssetSpecification:
    properties:
      asset_id:
        type: integer
      asset_name:
        description: Transient
        type: string
      asset_tag:
        type: string
      bios_version:
        description: Firmware & OS
        type: string
      color:
        type: string
      cpu_cores:
        type: integer
      cpu_model:
        description: CPU
        type: string
      cpu_speed_ghz:
        type: number
      created_at:
        type: string
      created_by:
        type: integer
      firmware_version:
        type: string
      form_factor:
        description: Physical
        type: string
      id:
        type: integer
      ip_address:
        type: string
      last_scanned_at:
        description: Audit
        type: string
      mac_address:
        description: Network
        type: string
      os_license_key:
        type: string
      os_name:
        type: string
      os_version:
        type: string
      ram_gb:
        description: Memory
        type: number
      ram_type:
        type: string
      resolution:
        type: string
      screen_size_inch:
        description: Display
        type: number
      storage_gb:
        description: Storage
        type: number
      storage_type:
        type: string
      updated_at:
        type: string
      weight_kg:
        type: number
    type: object
...
info:
  contact: {}
  description: Backend API implementing ISO/IEC 19770-10 Grade A++ RBAC model
  title: IT Asset & Compliance API
  version: "1.1"
paths:
...
  /assets/{id}/disposal:
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Ambil record disposal suatu aset
      tags:
      - Asset Disposal
    post:
      consumes:
      - application/json
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Buat disposal record untuk aset yang akan/sudah dibuang
      tags:
      - Asset Disposal
    put:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: Update disposal record (sertifikat, environmental compliance, dll)
      tags:
      - Asset Disposal
  /assets/{id}/spec:
    get:
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AssetSpecification'
      summary: Ambil spesifikasi hardware suatu aset
      tags:
      - Asset Specifications
    put:
      consumes:
      - application/json
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Buat atau update spesifikasi hardware aset (upsert)
      tags:
      - Asset Specifications
...
  /assets/disposals:
    get:
      parameters:
      - description: Filter yang sudah environmental compliant
        in: query
        name: env_compliant
        type: boolean
      produces:
      - application/json
      responses: {}
      summary: List semua disposal records
      tags:
      - Asset Disposal
...
  /licenses/{id}/usage:
    get:
      parameters:
      - description: License ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: List usage logs untuk suatu lisensi
      tags:
      - Software Usage
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      summary: Catat sesi penggunaan lisensi (manual / import)
      tags:
      - Software Usage
  /licenses/reconciliation:
    get:
      parameters:
      - description: 'Filter: compliant|over_licensed|under_utilized'
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses: {}
      summary: Laporan rekonsiliasi lisensi (entitlement vs aktual)
      tags:
      - Software Usage
...
swagger: "2.0"